requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
pydantic
python-dotenv
passlib[argon2]
PyJWT
